refactored proxy tree
